test/requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.11
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
fonttools==4.50.0
idna==3.4
importlib_resources==6.4.0
oauthlib==3.2.2
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pyparsing==3.1.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.4
zipp==3.18.1